// frontend/src/api/generated/models/updateJobInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Palmimo Portal
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateJobInfo {
  error: string | null;
  finished_at: number | null;
  kind: string;
  restarting_at: number | null;
  started_at: number | null;
  state: string;
  step: string | null;
  target: string | null;
}